+++ CHANGELOG.md
@@ -41,6 +41,7 @@ _This sections lists changes committed since most recent release_
 - Unit tests for IncludeRequiredFileMsg and GraphicUpload
 - A TeamDropdown component
 - Add getFileNameFromUrl utility function
+- Internal visibility for Graphic projects
 
 **Changed:**
 
@@ -89,11 +90,13 @@ _This sections lists changes committed since most recent release_
 - Rename to IncludeRequiredFileMsg from IncludeVideoFileMsg for reusability
 - Use TeamDropdown instead of an input field in ProjectDetailsForm for graphic projects
 - Connect team.id in the graphic buildFormTree function to allow graphic users to change a graphic project's team
-- Adjust the graphic project team queries to handle GPA Editorial & Design or ShareAmerica, i.e., the two TeamDropdown options for graphic projects
+- Adjust the graphic project team queries to handle GPA Editorial & Design or ShareAmerica, i.e., the two TeamDropdown options for graphic 
+- Use updateGraphicProject mutation to handle deletion of graphic support files
+- Update test data for GraphicSupportFiles to use updateGraphicProject
 - Remove the Social Media Share functionality for Graphics 
 - Use API URL to download video caption files 
 - Display the ShareAmerica logo in the graphic project preview and graphic project results card if the project's owner is ShareAmerica
- 
+
  
 
 **Fixed:**